requirements.txt
PyJWT
requests
cryptography
pyinstaller